# Mooncake store get zero copy meta optimization