@huangtingwei9988 2025-09-01
> I found that the impact of this optimization varies significantly across different GPUs. On the A10, the performance difference is not significant. However, on the H800, when len(indices) = 8192, this optimization reduces the execution time of a single `get_buffer_meta` call from 0.5 to 3 seconds down to just 1 to 2 ms.

I was surprised to see such an improvement on the H800！！